requirements.txt
camelot-py[all]
pandas
pdf2docx
